Moved: sysdiagnosis funcs out of ilapfuncs
